capsules: move virtual_kv to virtualizers/
