+++ capsules/extra/src/lib.rs
@@ -115,5 +115,4 @@ pub mod touch;
 pub mod tsl2561;
 pub mod usb;
 pub mod usb_hid_driver;
-pub mod virtual_kv;
 pub mod virtualizers;
